add app config setting for blackout date deletion

+++ lib/tasks/delete_old_blackouts.rake
@@ -1,7 +1,11 @@
 desc "Delete old blackouts"
 task :delete_old_blackouts => :environment do
-  #remove all blackouts older than the current day
-  old_blackouts = Blackout.where("end_date < ?", Date.today - 1.month)
+  # remove all blackouts older than the current day
+  # do not delete if appconfig isn't set up or if
+  # blackout exp date is not set
+  return if AppConfig.blank? || AppConfig.first.blackout_exp_time.blank?
+  time = AppConfig.first.blackout_exp_time
+  old_blackouts = Blackout.where("end_date < ?", Date.today - time.day)
   old_blackouts.each do |b|
     Rails.logger.info "Deleting old blackout:\n #{b.inspect}"
     b.destroy
